requirements.txt
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
looker-sdk==24.16.2
python-dotenv==1.0.1